requirements.txt
beautifulsoup4==4.8.0
Django==2.2.1
django-bootstrap4==1.0.1
django-chroniker==1.0.16
django-cors-headers==3.0.1
django-dbbackup==3.2.0
django-enumchoicefield==1.1.0
django-js-asset==1.2.2
django-location-field==2.1.0
django-mptt==0.10.0
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
drf-nested-routers==0.91
numpy==1.17.3
Pillow==6.0.0
psutil==5.6.3
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2019.1
scipy==1.3.1
six==1.12.0
soupsieve==1.9.3
sqlparse==0.3.0
toposort==1.5
